add transactional outbox pattern

+++ score-maker/src/utils.py
@@ -1,11 +1,13 @@
 import aiohttp
+import json
 from fastapi.exceptions import HTTPException
 from dateutil import parser
 from datetime import datetime, timezone
 from sqlalchemy import select
 from .database import async_session_maker
-from .models import ScoreModel
+from .models import ScoreModel, OutboxMessageModel
 from .config import LP_URL, LP_PORT
+from .schemas import KafkaEvent
 from sqlalchemy.ext.asyncio import AsyncSession
 
 
@@ -62,12 +64,36 @@ class ScoreDBMethods:
 
     @staticmethod
     async def insert_into_score(event_id: int, score: int):
-        new_score = ScoreModel(event_id=event_id, score=score)
         async with await ScoreDBMethods._get_session() as session:
-            session.add(new_score)
-            await session.commit()
-            await session.refresh(new_score)
-            return new_score.id
+            try:
+                new_score = ScoreModel(event_id=event_id, score=score)
+                session.add(new_score)
+                await session.commit()
+                await session.refresh(new_score)
+                row_id = new_score.id
+
+                data = KafkaEvent(
+                    data={
+                        "event_id": event_id,
+                        "event_score": score,
+                        "row_id": row_id
+                    },
+                    event="score_insert_into_db",
+                    status="success"
+                )
+
+                new_event = OutboxMessageModel(
+                    occurred_on=datetime.now(timezone.utc),
+                    status="в процессе",
+                    data=json.dumps(data.dict())
+                )
+                session.add(new_event)
+                await session.commit()
+                return {"status": "success"}
+
+            except Exception as e:
+                await session.rollback()
+                return {"status": "failed", "error": e}
 
     @staticmethod
     async def delete_from_score(row_id: int):
@@ -76,7 +102,12 @@ class ScoreDBMethods:
                 select(ScoreModel)
                 .where(ScoreModel.id == row_id)
             )
-            score_row = result.scalar_one()
+            score_row = result.scalar_one_or_none()
+
+            # Проверяем, что запись найдена
+            if score_row is None:
+                print(f"No score found with id: {row_id}")
+                return
             await session.delete(score_row)
             await session.commit()
 
@@ -85,7 +116,9 @@ class EventValidator:
     @staticmethod
     def validate_deadline(deadline):
         if deadline < datetime.now(timezone.utc):
-            raise HTTPException(status_code=400, detail="Событие нельзя оценить")
+            raise HTTPException(
+                status_code=400,
+                detail="Событие нельзя оценить")
 
 
 class ScoreMethods:
@@ -98,4 +131,3 @@ class ScoreMethods:
             if deadline > datetime.now(timezone.utc):
                 returned_events.append(event)
         return returned_events
-
